Fix and clean loadRelationshipsForMany
The supposed "array cleaning" by
```javascript
for (var i in recordsArray) {
  if (recordsArray.hasOwnProperty(i)) {
    recordsToBeLoaded.push(recordsArray[i]);
  }
}
```
did actually introduce two extra spurious entries in a case of mine,
so I prefer to rely on the native forEach function to make the right
call of what to include in the iteration.

Also replace the recursive function loadNextRecord by a more functional
variant akin to a left fold.